+++ code/src/nuvla/connector/utils.py
@@ -3,8 +3,8 @@
 import os
 import hashlib
 from datetime import datetime, timedelta
-from subprocess import run, PIPE, STDOUT
-from tempfile import NamedTemporaryFile, TemporaryDirectory
+from subprocess import run, PIPE, STDOUT, TimeoutExpired
+from tempfile import NamedTemporaryFile
 
 
 def _time_rm_nanos(time_str):
@@ -44,7 +44,12 @@ def append_os_env(env):
 def execute_cmd(cmd, **kwargs):
     opt_env = append_os_env(kwargs.get('env'))
     opt_input = kwargs.get('input')
-    result = run(cmd, stdout=PIPE, stderr=STDOUT, env=opt_env, input=opt_input, encoding='UTF-8')
+    timeout = kwargs.get('timeout', 120)
+    try:
+        result = run(cmd, stdout=PIPE, stderr=STDOUT, env=opt_env, input=opt_input,
+                     timeout=timeout, encoding='UTF-8')
+    except TimeoutExpired:
+        raise Exception('Command execution timed out after {} seconds'.format(timeout))
     if result.returncode == 0:
         return result.stdout
     else: